backend/requirements.txt
fastapi[standard]
uvicorn
passlib[bcrypt]
sqlalchemy
python-jose[cryptography]
boto3
python-multipart
psycopg2-binary
paypalrestsdk
stripe